TR-1345 - Emails send and close the modal
