+++ src/pages/templatesV2/EditAndPreviewPage.container.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import { connect } from 'react-redux';
+import { showAlert } from 'src/actions/globalAlert';
 import { getDraft, getPreview, getPublished, update as updateDraft, publish as publishDraft, deleteTemplate, create as createTemplate, sendPreview } from 'src/actions/templates';
 import { list as listDomains } from 'src/actions/sendingDomains';
 import { list as listSubaccounts } from 'src/actions/subaccounts';
@@ -55,7 +56,8 @@ const mapDispatchToProps = {
   publishDraft,
   listDomains,
   listSubaccounts,
-  sendPreview
+  sendPreview,
+  showAlert
 };
 
 export default connect(mapStateToProps, mapDispatchToProps)(EditAndPreviewPageContainer);
